requirements.txt
asgiref==3.8.1
cffi==1.17.0
click==8.1.7
coverage==7.6.1
cryptography==43.0.0
Django==5.1
djangorestframework==3.15.2
gunicorn==23.0.0
h11==0.14.0
jwt==1.3.1
numpy==2.1.0
packaging==24.1
pandas==2.2.2
pycparser==2.22
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
uvicorn==0.30.6
